Uploaded Zombie Survival Game
Run this on repl.it as it imports REPLIT in the script. I could make a better game but this a year old and shows my decent-ish programming skills from back then so I though you guys might want to see this Its honestly kind of fun if not bugged.